+++ provision/whmcs.sh
@@ -21,16 +21,8 @@ install_whmcs()
 
 configure_whmcs_nginx()
 {
-        local _datadir="$ZFS_DATA_MNT/whmcs"
-        if [ -f "$_datadir/etc/nginx-locations.conf" ]; then
-            tell_status "preserving $_datadir/etc/nginx-locations.conf"
-            return
-        fi
+    configure_nginx_server_d "whmcs" <<'EO_NGINX_WHMCS'
 
-        tell_status "saving $_datadir/etc/nginx-locations.conf"
-        tee "$_datadir/etc/nginx-locations.conf" <<'EO_NGINX_WHMCS'
-
-        listen       80;
         server_name  theartfarm.com www.theartfarm.com;
 
         if ($request_method !~ ^(GET|HEAD|POST)$ ) {
@@ -70,11 +62,6 @@ configure_whmcs_nginx()
         error_page  404              /404.html;
         location = /404.html {
             root   /data/html/;
-        }
-
-        error_page   500 502 503 504  /50x.html;
-        location = /50x.html {
-            root   /usr/local/www/nginx-dist;
         }
 
 EO_NGINX_WHMCS